% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_prices.R
\name{.as_year}
\alias{.as_year}
\title{Year of a time vector}
\usage{
.as_year(x)
}
\arguments{
\item{x}{A vector of dates or years.}
}
\value{
An integer vector of years.
}
\description{
Accepts the \code{Date} columns used throughout the package as well as plain
numeric, character or factor years.
}
\keyword{internal}
