% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sources.R
\name{.growth_cor}
\alias{.growth_cor}
\title{Correlation of yearly relative changes}
\usage{
.growth_cor(x, y)
}
\arguments{
\item{x, y}{Numeric vectors ordered by time.}
}
\description{
Correlation of yearly relative changes
}
\keyword{internal}
