% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{.vintage_filename}
\alias{.vintage_filename}
\title{File name of a vintage copy}
\usage{
.vintage_filename(filename, vintage)
}
\arguments{
\item{filename}{Base file name.}

\item{vintage}{Vintage label, e.g. a year.}
}
\description{
File name of a vintage copy
}
\keyword{internal}
