% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_ea}
\alias{geo_ea}
\title{Euro area countries}
\format{
A character vector of length 20.
}
\source{
data-raw/data_main.R
}
\usage{
geo_ea
}
\description{
The twenty countries of the euro area, as Eurostat two letter codes (\code{EL} for
Greece, not \code{GR}).
}
\keyword{datasets}
