% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_names_list}
\alias{geo_names_list}
\title{Country names in Finnish and English}
\format{
A list of two named character vectors, \code{fi} and \code{en}, each named by
Eurostat country code.
}
\source{
data-raw/naming.R
}
\usage{
geo_names_list
}
\description{
Names for the countries the reports use, for labelling figures and tables.
}
\seealso{
\code{\link[=geo_names]{geo_names()}}
}
\keyword{datasets}
