% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase.R
\name{rebase_index}
\alias{rebase_index}
\title{Rebase a numeric series to an index with a chosen base year}
\usage{
rebase_index(x, time, baseyear, basevalue = 100)
}
\arguments{
\item{x}{Numeric vector of values.}

\item{time}{A vector of dates or years corresponding to \code{x}.}

\item{baseyear}{Numeric year or vector of years to use as base.}

\item{basevalue}{Numeric index value assigned to the base period.
If \code{NULL}, the base value becomes the mean of the base-year observations.}
}
\value{
A numeric vector of rebased index values.
}
\description{
This function converts a numeric time series into an index, where the level
in the specified \code{baseyear} equals \code{basevalue}. The base-year value is the
mean of all observations in that year. Works in both base R and inside
\code{dplyr::mutate()}.
}
\examples{
df |> mutate(index = rebase_index(values, time, 2007))
df |> mutate(index = rebase_index(values, time, baseyear = c(2006, 2007)))

}
