% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightning.R
\name{weight_all}
\alias{weight_all}
\alias{weight_at}
\title{Mass weighting}
\usage{
weight_all(.data, geo, time, except, weight_df)

weight_at(.data, geo, time, at, weight_df)
}
\arguments{
\item{.data}{A tbl.}

\item{geo}{A vector to indicate countries.}

\item{time}{A year (or date which is converted to a year) for weights.}

\item{except}{Variables to exclude, a character vector.}

\item{ar}{Variables to include, a character vector.}

\item{weitht_df}{A weigthing data.frame in long form. Should have time, geo_base and geo columns.}
}
\description{
Weight all variables except.
}
\section{Functions}{
\itemize{
\item \code{weight_at()}: To spesify variables to weight

}}
