% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightning.R
\name{weight_function}
\alias{weight_function}
\title{Weight function for weight_index}
\usage{
weight_function(one_geo, x, geo, w_df, check_geos, mean_type)
}
\arguments{
\item{one_geo}{A character of base geo}

\item{x}{A variable to weight}

\item{geo}{A vector of ids to weights}

\item{w_df}{A weigth data.frame}

\item{check_geos}{A locigal to check if all geos are in w_df}

\item{mean_type}{A meant type to calculate "geom" for geometric and "arit" for arithmetic}
}
\description{
Weight function for weight_index
}
\keyword{internal}
